Change chain_index and pos_in_chain to u32

+++ src/inc_encoding.rs
@@ -41,7 +41,7 @@ pub trait IncomparableEncoding {
         message: &[u8; MESSAGE_LENGTH],
         randomness: &Self::Randomness,
         epoch: u32,
-    ) -> Result<Vec<u64>, EncodingError>;
+    ) -> Result<Vec<u32>, EncodingError>;
 }
 
 pub mod basic_winternitz;
